add decimal places to price field

+++ catalog/models.py
@@ -81,7 +81,7 @@ class MedicineProduct(models.Model):
         max_length=100, choices=BUY_PLACES, blank=True, null=True
     )
     product_url = models.CharField(max_length=255, null=True, blank=True)
-    price = models.IntegerField(null=True, blank=True)
+    price = models.DecimalField(max_digits=10, decimal_places=2, null=True, blank=True)
     currency = models.CharField(max_length=10,choices=CURRENCY_CHOICES, default="RON", null=True, blank=True)
     notes = models.TextField(null=True, blank=True)
     method_of_administration = models.TextField(null=True, blank=True)
